requirements.txt
asgiref==3.4.1
certifi==2021.5.30
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.6
django-stdimage==5.3.0
gunicorn==20.1.0
Pillow==8.3.1
psycopg2-binary==2.9.1
pytz==2021.1
sqlparse==0.4.1
static3==0.7.0
typing-extensions==3.10.0.0
wincertstore==0.2
